Cambiado el tipo de las cooredenadas de Vector para evitar perdidas de precisión en el redondeo, ahora son "double". Adaptado el resto del código.
git-svn-id: http://pong-opengl.googlecode.com/svn/trunk@24 841c6515-18fd-ad39-5d5d-772e54db8338

+++ tipos/include/Vector.h
@@ -4,8 +4,8 @@
 #define _VECTOR
 
 typedef struct {
-	int x;
-	int y;
+	double x;
+	double y;
 }Vector; 
 
 const Vector operator+(const Vector v1, const Vector v2);
@@ -13,7 +13,7 @@ const Vector operator-(const Vector v1, const Vector v2);
 const Vector operator*(const int esc, const Vector v);
 const Vector operator*(const Vector v, const int esc);
 const Vector operator/(const Vector v, const int esc);
-const int operator*(const Vector v1, const Vector v2);
+const double operator*(const Vector v1, const Vector v2);
 const Vector operator+=(Vector& v1, const Vector v2);
 const Vector operator-=(Vector& v1, const Vector v2);
 const Vector operator*=(Vector& v1, const int esc);
